requirements.txt
flask==2.3.3
mfrc522==0.0.7
RPi.GPIO==0.7.1
